[#214] Rearranged some files
