+++ Server/src/tests/models/DatasetQueryModel.spec.ts
@@ -1,6 +1,5 @@
 import { createConnection, getConnection } from 'typeorm';
-
-import { DataQueryModel } from '../../models/DatasetQueryModel';
+import { DataQueryModel } from '../../models/DatasetModels/DatasetQueryModel';
 
 describe('data service test', () => {
     let dataQueryModel: DataQueryModel;
